requirements.txt
apache-airflow
apache-airflow-providers-mysql
apache-airflow-providers-postgres
pandas
mysql-connector-python
sqlalchemy
psycopg2-binary
